Move ReactComponentFactoryProxy to react.dart
